Moving methods to Adapter
