+++ OneClickLCA_Adapter/Objects/MaterialsCarbonResourceTypeWrappers.cs
@@ -1,0 +1,13 @@
+using BH.oM.Adapters.OneClickLCA;
+
+namespace BH.Adapter.OneClickLCA.Objects
+{
+    // Adapter-side wrapper types that reuse the oM definitions.
+    public class MaterialsCarbonResourceIdentifiers : BH.oM.Adapters.OneClickLCA.MaterialsCarbonResourceIdentifiers { }
+    public class MaterialsCarbonResourceNaming : BH.oM.Adapters.OneClickLCA.MaterialsCarbonResourceNaming { }
+    public class MaterialsCarbonResourceDataSource : BH.oM.Adapters.OneClickLCA.MaterialsCarbonResourceDataSource { }
+    public class MaterialsCarbonResourcePhysical : BH.oM.Adapters.OneClickLCA.MaterialsCarbonResourcePhysical { }
+    public class MaterialsCarbonResourceOrganisation : BH.oM.Adapters.OneClickLCA.MaterialsCarbonResourceOrganisation { }
+    public class MaterialsCarbonImpactModule : BH.oM.Adapters.OneClickLCA.MaterialsCarbonImpactModule { }
+    public class MaterialsCarbonResourceDeclaredImpacts : BH.oM.Adapters.OneClickLCA.MaterialsCarbonResourceDeclaredImpacts { }
+}
